Changed some text styling on overviewpage and moved overviewpage from coontainers to components
